Change example query type
The example with ContainerType for me throws an error:

 "A binary operator with incompatible types was detected. Found operand types 'microsoft.graph.containerType' and 'Edm.String' for operator kind 'Equal'.",

And when you look at whats returned from Trending it is media type that has the application/vnd.openxmlformats-officedocument.wordprocessingml.document, not container type. If I switch the example to use MediaType it works. So thought I'd propose this as a change. So just to be cleared this query works for me on media type:

https://graph.microsoft.com/beta/me/insights/trending?$filter=ResourceVisualization/MediaType eq 'application/vnd.openxmlformats-officedocument.wordprocessingml.document'

Where as the example one threw an error for me

https://graph.microsoft.com/beta/me/insights/trending?$filter=ResourceVisualization/ContainerType eq 'application/vnd.openxmlformats-officedocument.wordprocessingml.document'