Clarify the contents of the Response from the API
When using the API, it wasn't clear right away that the `value` property that was returned was a Base64 encoded file, nor what I needed to do to it in order to use it with devices. While I quickly recognized it as Base64, I initially tried converting from Base64 directly to UTF-8 encoded text, which looked right when compared to a manually-downloaded `.mobileconfig` file from Intune, but didn't work when tested with iOS devices. I figure that clarifying this here might save other people from some headaches.